+++ lib/overcommit/hook_context/run_all.rb
@@ -6,16 +6,18 @@ module Overcommit::HookContext
   # This results in pre-commit hooks running against the entire repository,
   # which is useful for automated CI scripts.
   class RunAll < Base
-    EMPTY_SET = Set.new
-
     def modified_files
       all_files
     end
 
-    # Return an empty set since in this context the user didn't actually touch
-    # any lines.
-    def modified_lines_in_file(_file)
-      EMPTY_SET
+    # Returns all lines in the file since in this context the entire repo is
+    # being scrutinized.
+    #
+    # @param file [String]
+    # @return [Set]
+    def modified_lines_in_file(file)
+      @modified_lines_in_file ||= {}
+      @modified_lines_in_file[file] ||= Set.new(1..(count_lines(file) + 1))
     end
 
     def hook_class_name
@@ -34,6 +36,11 @@ module Overcommit::HookContext
 
     def all_files
       @all_files ||= Overcommit::GitRepo.all_files
+    end
+
+    def count_lines(file)
+      result = Overcommit::Utils.execute(%w[wc -l] + [file])
+      result.success? ? result.stdout.to_i : 0
     end
   end
 end